BLE: Warning cleanup
- Missing default assignment operator
- Field order in construction list
- Typo
- Unused header

+++ connectivity/FEATURE_BLE/include/ble/common/ble/gap/AdvertisingDataParser.h
@@ -70,7 +70,7 @@ public:
      */
     bool hasNext() const
     {
-        if (position >= data.size()) {
+        if (position >= (size_t) data.size()) {
             return false;
         }
 
@@ -79,7 +79,7 @@ public:
             return false;
         }
 
-        if (position + current_length() >= data.size()) {
+        if (position + current_length() >= (size_t) data.size()) {
             return false;
         }
